requirements.txt
Django==4.2
python-dotenv==1.0.0
django-smtp-ssl==1.0
django-cookie-law==2.2.0
six==1.16.0